Lists of paths allowed in `captus extract`
- You can use a list of paths to individual sample assembly folders (ending in __captus-asm) to `captus extract`. Useful when your assemblies are in separate captus assembly folders
- The name in symbolic links are honored
- Reduced default `--min_coverage` to 0.2 in `captus align`

+++ captus/new_targets_from_alignments.py
@@ -603,7 +603,7 @@ def select_targets(
 
         header_pieces = cluster[0].split()
         seq_name = header_pieces[0].replace(">", "")
-        num_seqs = len(cluster) / 2
+        num_seqs = int(len(cluster) / 2)
         samples = sorted(list(set(samples)))
         num_samples = len(samples)
